calypso_AgentiArena/scribe-agent/requirements.txt
fastapi
uvicorn
langgraph
pydantic
requests
google-generativeai
beautifulsoup4
